Tab space & error visualization
